+++ large-records-benchmarks/bench/after/After/Sized/R070.hs
@@ -23,7 +23,6 @@ module After.Sized.R070 where
 
 import Data.Aeson (ToJSON(..))
 import Data.Record.Generic.JSON
-import Data.Record.Plugin
 
 import Bench.Types