+++ large-records/test/Test/Record/Sanity/QualifiedImports/A.hs
@@ -12,7 +12,5 @@
 
 module Test.Record.Sanity.QualifiedImports.A (T(..)) where
 
-import Data.Record.Plugin
-
 {-# ANN type T largeRecord #-}
 data T a = MkT { x :: Int, y :: [a] }