+++ large-records/test/Test/Record/Sanity/RDP/SingleModule.hs
@@ -20,8 +20,6 @@ module Test.Record.Sanity.RDP.SingleModule (tests) where
 import Test.Tasty
 import Test.Tasty.HUnit
 
-import Data.Record.Plugin
-
 {-------------------------------------------------------------------------------
   Simple field selection and override
 -------------------------------------------------------------------------------}